Move classes to different packages
