+++ ledger/ledger-utxo/src/test/java/net/corda/v5/ledger/utxo/AbstractMockTestHarness.java
@@ -5,8 +5,7 @@ import net.corda.v5.application.crypto.DigitalSignatureMetadata;
 import net.corda.v5.base.types.MemberX500Name;
 import net.corda.v5.crypto.DigitalSignature;
 import net.corda.v5.crypto.SecureHash;
-import net.corda.v5.ledger.common.transaction.Party;
-import net.corda.v5.ledger.utxo.transaction.UtxoLedgerService;
+import net.corda.v5.ledger.common.Party;
 import net.corda.v5.ledger.utxo.transaction.UtxoLedgerTransaction;
 import net.corda.v5.ledger.utxo.transaction.UtxoSignedTransaction;
 import net.corda.v5.ledger.utxo.transaction.UtxoTransactionBuilder;
